requirements.txt
fdk
requests
python-telegram-bot
oci
